% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{boxplot_generator}
\alias{boxplot_generator}
\title{Generate a plotly boxplot}
\usage{
boxplot_generator(
  calr_summary,
  calr_circadian,
  parameter,
  axis_text,
  color_list
)
}
\arguments{
\item{calr_summary}{a calr summary}

\item{calr_circadian}{a circadian calr summary}

\item{parameter}{parameter to display (ee, eb, feed, vo2)}

\item{axis_text}{text to display on second axis}

\item{color_list}{a data frame with group and colors}
}
\value{
a plotly object
}
\description{
Generate a plotly boxplot
}
