% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{calculate_res}
\alias{calculate_res}
\title{calculate resolution}
\usage{
calculate_res(data_set)
}
\arguments{
\item{data_set}{}
}
\value{
calculated resolution
}
\description{
calculate resolution
}
