% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{circadian_summary}
\alias{circadian_summary}
\title{Circadian summary}
\usage{
circadian_summary(trimmed_calr, resolution)
}
\arguments{
\item{trimmed_calr}{a trimmed calR object}
}
\value{
a circadian average of values
}
\description{
Circadian summary
}
