% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{convert_to_calr}
\alias{convert_to_calr}
\title{Convert to CalR}
\usage{
convert_to_calr(calr_file, calr_headers)
}
\arguments{
\item{calr_file}{a csv file generated by the sable macro}

\item{calr_headers}{a list of headers found in the data output from CalR}
}
\value{
a data file in the CalR format
}
\description{
Convert to CalR
}
