% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{group_assigner}
\alias{group_assigner}
\title{Group assigner}
\usage{
group_assigner(session_file)
}
\arguments{
\item{session_file}{a calR session file}
}
\value{
a data frame showing group distribution
}
\description{
Group assigner
}
