% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{pull_colors}
\alias{pull_colors}
\title{Pull colors from session file}
\usage{
pull_colors(session_file)
}
\arguments{
\item{session_file}{from calR}
}
\value{
a data frame with group name and colors from calR
}
\description{
Pull colors from session file
}
