% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{trim_calr_data}
\alias{trim_calr_data}
\title{Process calR data}
\usage{
trim_calr_data(calr, session, group_info, resolution)
}
\arguments{
\item{calr}{a calR file in the calR format}

\item{session}{a session file}

\item{group_info}{a dataframe containing info on group distributions}

\item{resolution}{resolution of dataset}
}
\value{
a trimmed calR object
}
\description{
Process calR data
}
