% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{generate_formula}
\alias{generate_formula}
\title{Generate model formula}
\usage{
generate_formula(parameter)
}
\arguments{
\item{parameter}{a parameter as a text string from a set list}
}
\value{
a formula for modelling
}
\description{
Generate model formula
}
