% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{generate_summary_data}
\alias{generate_summary_data}
\title{Generate summary info}
\usage{
generate_summary_data(calr, session, group_info, resolution)
}
\arguments{
\item{calr}{a calR file in the calR format}

\item{session}{a session file}

\item{group_info}{a dataframe containing info on group distributions}

\item{resolution}{calculated resolution of the dataset}
}
\value{
mean and SEM for various parameters
}
\description{
Generate summary info
}
