% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{xy_plotter}
\alias{xy_plotter}
\title{XY plotter}
\usage{
xy_plotter(summary_calr, color_key)
}
\arguments{
\item{summary_calr}{a calr summary object}

\item{color_key}{a dataframe with group names and color}
}
\value{
a plotly object
}
\description{
XY plotter
}
